+++ tool/tctl/common/collection.go
@@ -1135,43 +1135,6 @@ func (c *vnetConfigCollection) WriteText(w io.Writer, verbose bool) error {
 	return trace.Wrap(err)
 }
 
-type accessRequestCollection struct {
-	accessRequests []types.AccessRequest
-}
-
-func (c *accessRequestCollection) Resources() []types.Resource {
-	r := make([]types.Resource, len(c.accessRequests))
-	for i, resource := range c.accessRequests {
-		r[i] = resource
-	}
-	return r
-}
-
-func (c *accessRequestCollection) WriteText(w io.Writer, verbose bool) error {
-	var t asciitable.Table
-	var rows [][]string
-	for _, al := range c.accessRequests {
-		var annotations []string
-		for k, v := range al.GetSystemAnnotations() {
-			annotations = append(annotations, fmt.Sprintf("%s/%s", k, strings.Join(v, ",")))
-		}
-		rows = append(rows, []string{
-			al.GetName(),
-			al.GetUser(),
-			strings.Join(al.GetRoles(), ", "),
-			strings.Join(annotations, ", "),
-		})
-	}
-	if verbose {
-		t = asciitable.MakeTable([]string{"Name", "User", "Roles", "Annotations"}, rows...)
-	} else {
-		t = asciitable.MakeTableWithTruncatedColumn([]string{"Name", "User", "Roles", "Annotations"}, rows, "Annotations")
-	}
-
-	_, err := t.AsBuffer().WriteTo(w)
-	return trace.Wrap(err)
-}
-
 type pluginCollection struct {
 	plugins []types.Plugin
 }
